chore: move pytorch_dataset to models folder
