+++ src/spira_training/shared/adapters/pytorch/model_trainer/implementations/simple_pytorch_dataloader.py
@@ -1,10 +1,13 @@
 from typing import Literal, Sequence
 
+from src.spira_training.shared.adapters.pytorch.models.pytorch_dataset import (
+    PytorchDataset,
+)
+
 from src.spira_training.shared.adapters.pytorch.models.pytorch_batch import (
     PytorchBatch,
 )
 
-from .pytorch_dataset import PytorchDataset
 from src.spira_training.shared.adapters.pytorch.model_trainer.interfaces.pytorch_dataloader import (
     PytorchDataloader,
 )
